+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Scroll on affiliates
+- Send all affiliates to export orders via new variable in the query
+
 ## [1.2.1] - 2022-09-22
 
 ### Added